+++ libraries/docker_service_manager_systemd.rb
@@ -23,7 +23,7 @@ module DockerCookbook
           docker_socket: connect_socket
         )
         action connect_socket.nil? ? :delete : :create
-        not_if { docker_name == 'default' && ::File.exist?('/lib/systemd/system/docker.socket') }
+        not_if { docker_name == 'docker' && ::File.exist?('/lib/systemd/system/docker.socket') }
       end
 
       directory '/etc/containerd'
@@ -59,7 +59,7 @@ module DockerCookbook
           docker_socket: connect_socket,
           containerd: docker_containerd
         )
-        not_if { docker_name == 'default' && ::File.exist?('/lib/systemd/system/docker.service') }
+        not_if { docker_name == 'docker' && ::File.exist?('/lib/systemd/system/docker.service') }
       end
 
       # this overrides the main systemd socket
@@ -128,6 +128,10 @@ module DockerCookbook
         supports status: true
         action [:disable, :stop]
         only_if { ::File.exist?("/lib/systemd/system/#{docker_name}.service") }
+      end
+
+      systemd_unit "#{docker_name}.socket" do
+        action [:disable, :stop]
       end
     end