+++ test/cookbooks/docker_test/recipes/install_and_stop.rb
@@ -1,0 +1,3 @@
+docker_service 'default' do
+  action [:create, :stop]
+end