+++ test/integration/smoke/inspec/assert_functioning_spec.rb
@@ -21,3 +21,11 @@ describe command('docker --host=unix:///var/run/docker-one.sock ps -a') do
   its(:stdout) { should match(/hello-world/) }
   its(:stdout) { should_not match(/an_echo_server/) }
 end
+
+if os.family == 'redhat' && os.release.to_i < 8
+  # verify if docker-ce install was not changed
+  describe command("rpm -V docker-ce | grep '^..5'") do
+    its(:exit_status) { should eq 1 }
+    its(:stdout) { should eq '' }
+  end
+end